+++ src/main/java/com/skyflow/Configuration.java
@@ -6,7 +6,7 @@ import com.skyflow.entities.LogLevel;
 import com.skyflow.logs.InfoLogs;
 
 public final class Configuration {
-    public static void setLoggingLevel(LogLevel level) {
+    public static void setLogLevel(LogLevel level) {
         LogUtil.setupLogger(level);
         LogUtil.printInfoLog(Helpers.parameterizedString(InfoLogs.CurrentLogLevel.getLog(), String.valueOf(level)));
     }