+++ src/main/java/com/skyflow/vault/Skyflow.java
@@ -22,7 +22,7 @@ import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.FutureTask;
 
-public class Skyflow {
+public final class Skyflow {
     private final SkyflowConfiguration configuration;
 
     private Skyflow(SkyflowConfiguration config) {